+++ internal/cmd/flags.go
@@ -18,21 +18,39 @@ import (
 	"errors"
 
 	"github.com/sapcc/limesctl/v3/internal/core"
+	"github.com/spf13/cobra"
 )
+
+func doNotSortFlags(cmd *cobra.Command) {
+	cmd.Flags().SortFlags = false
+	cmd.LocalFlags().SortFlags = false
+	cmd.PersistentFlags().SortFlags = false
+	cmd.InheritedFlags().SortFlags = false
+}
 
 ///////////////////////////////////////////////////////////////////////////////
 // Limes' API request filter flags.
 
 type commonFilterFlags struct {
-	Services []string `sep:"," help:"Service type(s)."`
-	Areas    []string `sep:"," help:"Service area(s)."`
+	services []string
+	areas    []string
+}
+
+func (f *commonFilterFlags) AddToCmd(cmd *cobra.Command) {
+	cmd.Flags().StringSliceVar(&f.services, "services", nil, "service types (comma separated list)")
+	cmd.Flags().StringSliceVar(&f.areas, "areas", nil, "service areas (comma separated list)")
 }
 
 // resourceFilterFlags define parameters for Limes API requests that concern
 // resources.
 type resourceFilterFlags struct {
 	commonFilterFlags
-	Resources []string `sep:"," help:"Resource name(s)."`
+	resources []string
+}
+
+func (f *resourceFilterFlags) AddToCmd(cmd *cobra.Command) {
+	f.commonFilterFlags.AddToCmd(cmd)
+	cmd.Flags().StringSliceVar(&f.resources, "resources", nil, "resource names (comma separated list)")
 }
 
 // rateFilterFlags define parameters for Limes API requests that concern rates.
@@ -40,28 +58,38 @@ type rateFilterFlags struct {
 	commonFilterFlags
 }
 
+func (f *rateFilterFlags) AddToCmd(cmd *cobra.Command) {
+	f.commonFilterFlags.AddToCmd(cmd)
+}
+
 ///////////////////////////////////////////////////////////////////////////////
 // CLI output format flags.
 
 type commonOutputFmtFlags struct {
-	Format core.OutputFormat `short:"f" enum:"${outputFormats}" default:"table" help:"Output format (${enum})."`
-	Names  bool              `help:"Show output with names instead of UUIDs. Not valid for 'json' output format."`
-	Long   bool              `help:"Show detailed output. Not valid for 'json' output format."`
+	format core.OutputFormat
+	names  bool
+	long   bool
+}
+
+func (o *commonOutputFmtFlags) AddToCmd(cmd *cobra.Command) {
+	cmd.Flags().VarP(&o.format, "format", "f", "output format: table (default), json, csv")
+	cmd.Flags().BoolVar(&o.names, "names", false, "show output with names instead of UUIDs. Not valid for 'json' output format")
+	cmd.Flags().BoolVar(&o.long, "long", false, "show detailed output. Not valid for 'json' output format")
 }
 
 func (o commonOutputFmtFlags) validate() (*core.OutputOpts, error) {
 	// Catch errors.
-	if o.Long && o.Names {
-		return nil, errors.New("'--long' and '--names' flags are mutually exclusive")
+	if o.long && o.names {
+		return nil, errors.New("'--long' and '--names' flags are mutually exclusive, i.e. use one, not both")
 	}
 
 	opts := &core.OutputOpts{
-		Fmt: o.Format,
+		Fmt: o.format,
 	}
 	switch {
-	case o.Long:
+	case o.long:
 		opts.CSVRecFmt = core.CSVRecordFormatLong
-	case o.Names:
+	case o.names:
 		opts.CSVRecFmt = core.CSVRecordFormatNames
 	default:
 		opts.CSVRecFmt = core.CSVRecordFormatDefault
@@ -73,7 +101,12 @@ func (o commonOutputFmtFlags) validate() (*core.OutputOpts, error) {
 // resourceOutputFmtFlags define how the app will print resource data.
 type resourceOutputFmtFlags struct {
 	commonOutputFmtFlags
-	Humanize bool `help:"Show quota and usage values in an user friendly unit. Not valid for 'json' output format."`
+	humanize bool
+}
+
+func (o *resourceOutputFmtFlags) AddToCmd(cmd *cobra.Command) {
+	o.commonOutputFmtFlags.AddToCmd(cmd)
+	cmd.Flags().BoolVar(&o.humanize, "humanize", false, "show quota and usage values in an user friendly unit. Not valid for 'json' output format")
 }
 
 func (o resourceOutputFmtFlags) validate() (*core.OutputOpts, error) {
@@ -82,13 +115,17 @@ func (o resourceOutputFmtFlags) validate() (*core.OutputOpts, error) {
 		return nil, err
 	}
 
-	opts.Humanize = o.Humanize
+	opts.Humanize = o.humanize
 	return opts, nil
 }
 
 // rateOutputFmtFlags define how the app will print rate limit data.
 type rateOutputFmtFlags struct {
 	commonOutputFmtFlags
+}
+
+func (o *rateOutputFmtFlags) AddToCmd(cmd *cobra.Command) {
+	o.commonOutputFmtFlags.AddToCmd(cmd)
 }
 
 func (o rateOutputFmtFlags) validate() (*core.OutputOpts, error) {